test: split ProductsTest into per-domain form, action and table tests
